requirements.txt
gtts
openai
playsound
streamlit
streamlit-audiorecorder
sounddevice
#scipy
whisper